LOCAL: iommu: Add IOMMU device registry
Add an IOMMU device registry for drivers to register with and implement
a method for users of the IOMMU API to attach to an IOMMU device. This
allows to support deferred probing and gives the IOMMU API a convenient
hook to perform early initialization of a device if necessary.

Signed-off-by: Thierry Reding <treding@nvidia.com>

Conflicts:
	drivers/iommu/iommu.c
	include/linux/iommu.h