gpu: host1x: Map gathers to IOMMU domain
This patch adds a new IOMMU domain for host1x and makes the driver
map the CDMA pushbuffer and gathers in that domain. If the firewall
is enabled, only the gather copy buffer is mapped; otherwise, all
gathers are mapped. This allows disabling the firewall when IOMMU
support is enabled.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>