drm/tegra: Add job firewall
Add a firewall that validates jobs before submission to ensure
they don't do anything they aren't allowed to do, like accessing
memory they should not access.

The firewall is functionality-wise a copy of the firewall already
implemented in gpu/host1x. It is copied here as it makes more
sense for it to live on the DRM side, as it is only needed for
userspace job submissions, and generally the data it needs to
do its job is easier to access here.

In the future, the other implementation will be removed.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
---
v6:
* If .is_valid_class is not set, match against client's
  class
* Fix indentation
* Return to previous class after a WAIT command
* Print warning detailing error when rejecting a command
  buffer
* Removed uapi subdirectory
v5:
* Support SETCLASS opcode
v3:
* New patch