gpu: nova-core: load and patch Booter firmwares
The Booter load firmware is an essential part of booting the GSP on
Turing and Ampere.

Properly parse and patch the Booter load and unload firmware files,
storing them into a form that is ready to load into the sec2 falcon for
execution.

Booter load still cannot be run, as we need to prepare its parameters,
which includes the image of the GSP firmware to load.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>