scripts: Add rust_find_build_error.sh
Failed build_assert!() type assertions in Rust code cause the kernel
build to fail with a missing reference to the rust_build_error symbol,
with nothing to point to the failing assertion. As a result, debugging
such errors can be difficult.

It turns out that in such cases, the assertion message is still built
into the object file containing the asserting code. As build errors
by definition are resolved at build time, passing assertions don't
leave such data. As such, we can use tools like grep to find any
failing assertion messages to aid in debugging.

To that end, add a 'scripts/rust_find_build_error.sh' script that does
this. Also add the file ane line number of the assertion, as well as
the word 'build' to the assertion message to make them easier to find,
and easier to locate the assertion in the source code.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>